requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.2.0
dj-rest-auth[with_social]==5.0.1
djangorestframework-simplejwt==5.3.0            # for JWT
drf_spectacular==0.26.5
pycryptodome==3.19.0                            # for Cipher(Encryption/Decryption)
psycopg2==2.9.8                                 # for PostgreSQL
mysqlclient==2.2.0                              # for MySQL and MariaDB
cx-Oracle==8.3.0                                # for Oracle